+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.49] - 2026-04-29
+
+### Docs
+- Update README.md
+
+### Test
+- Update tests/test_doctor_cli.py
+- Update tests/test_doctor_orchestrator.py
+
+### Other
+- Update regres/doctor_cli.py
+- Update regres/doctor_orchestrator.py
+- Update regres/regres_cli.py
+
 ## [0.1.48] - 2026-04-29
 
 ### Docs